implement todo to add null check in delete method

+++ src/main/java/com/example/service/ProductServiceImpl.java
@@ -47,8 +47,12 @@ public class ProductServiceImpl implements ProductService {
                 }).orElse(null);
     }
 
-    public void delete(Long id){
+    public boolean delete(Long id){
+        if(!repo.existsById(id)) {
+            return false;
+        }
         repo.deleteById(id);
+        return true;
     }
 
 }
